requirements.txt
asgiref==3.8.1
cachetools==5.5.2
Django==5.2
django-rql==4.4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
flake8==7.2.0
lark-parser==0.11.0
lib-rql==2.0.2
mccabe==0.7.0
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
